chore: rename example files
